third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/suborigins/suborigin-service-worker-link-rel-expected.txt

CONSOLE ERROR: line 20: Cannot register service worker with <link> element. Service worker is disabled because the context is in a suborigin.
This is a testharness.js-based test.
PASS Service Worker link rel fails from a suborigin 
Harness: the test ran to completion.
